Issue #138: Option for JMM Server to change state to deleted rather than removing files from MyList when deleted from HDD
git-svn-id: http://jmm.googlecode.com/svn/trunk@1092 a6988de2-25aa-124f-8851-278695053f3e

+++ JMMServer/Commands/AniDB/CommandRequest_DeleteFileFromMyList.cs
@@ -60,12 +60,23 @@ namespace JMMServer.Commands
 
 			try
 			{
-				if (FileID > 0)
-					JMMService.AnidbProcessor.DeleteFileFromMyList(FileID);
-				else
-					JMMService.AnidbProcessor.DeleteFileFromMyList(Hash, FileSize);
+				if (ServerSettings.AniDB_MyList_DeleteType == AniDBAPI.AniDBFileDeleteType.Delete)
+				{
+					if (FileID > 0)
+						JMMService.AnidbProcessor.DeleteFileFromMyList(FileID);
+					else
+						JMMService.AnidbProcessor.DeleteFileFromMyList(Hash, FileSize);
 
-				logger.Info("Deleting file from list: {0}_{1}", Hash, FileID);
+					logger.Info("Deleting file from list: {0}_{1}", Hash, FileID);
+				}
+				else
+				{
+					if (FileID < 0)
+					{
+						JMMService.AnidbProcessor.MarkFileAsDeleted(Hash, FileSize);
+						logger.Info("Deleting file from list: {0}_{1}", Hash, FileID);
+					}
+				}
 			}
 			catch (Exception ex)
 			{
